Feat(#16) : 로그인 기능 구현 (#23)
* feat: 로그인 관련 DTO 추가

* feat: 직원 DTO 추가

* feat: 권한 DTO 추가, enum 추가

* fix: DTO 필드 정리 및 Lombok 어노테이션 분리

* feat: 직원 DAO 조회 메서드 구현

* fix: 디렉토리 경로 수정

# Conflicts:
#	src/main/java/auth/LoginResponseDTO.java
#	src/main/java/employee/EmployeeDTO.java
#	src/main/java/role/RoleDTO.java

* refactor: roleId int형으로 수정

* fix: 로그인 시 권한명 조회 가능하도록 DAO 수정

* refactor: 로그인 DTO와 직원 조회 DAO 구조 수정

- 기존 EmployeeDTO에서 JOIN 결과를 함께 처리하던 로직을 수정
- 로그인 응답 DTO와 직원 조회 DAO 역할을 분리

* refactor: roleId 기반 권한 타입 매핑 적용

* feat: Auth 관련 Service, Controller 구현

- 존재하지 않는 ID, 사용 중단 계정 검증

* refactor: 권한 타입 디렉토리 경로 변경

+++ src/main/java/employee/EmployeeDTO.java
@@ -21,7 +21,7 @@ public class EmployeeDTO {
   private String employeeName;
   private String phoneNumber;
 
-  private long roleId;
+  private int roleId;
   private Long storeId;
 
   private String isActive;
